Messy code, but it works somewhat

+++ api/services/caboodles.js
@@ -1,15 +1,16 @@
 const { ObjectId } = require("mongodb");
+const auth = require("../middleware/authenticateRequest");
 
-global.services["createCaboodle"] = async ({
-    name,
-    description,
-    user_id,
-    image,
-} = {}) => {
+global.services["createCaboodle"] = async (
+    { name, description, user_id, image },
+    header
+) => {
     try {
         if (!name || !description || !user_id)
             return { error: "Missing parameters" };
-
+        let authenticatedUser = auth.authenticateRequest(header);
+        if (user_id !== authenticatedUser.user_id)
+            return { error: "Unauthrized user" };
         const caboodle = {
             nickname: name,
             description,
@@ -26,9 +27,12 @@ global.services["createCaboodle"] = async ({
     }
 };
 
-global.services["LoadCaboodles"] = async ({ user_id, skip }) => {
+global.services["LoadCaboodles"] = async ({ user_id, skip }, header) => {
     try {
         if (!user_id) return { error: "Missing parameters" };
+        let authenticatedUser = auth.authenticateRequest(header);
+        if (user_id !== authenticatedUser.user_id)
+            return { error: "Unauthrized user" };
         const results = await global.database["Caboodles"]
             .find({ user_id: new ObjectId(user_id) })
             .skip(skip)
@@ -37,15 +41,20 @@ global.services["LoadCaboodles"] = async ({ user_id, skip }) => {
         const totalCaboodles = await global.database["Caboodles"]
             .find({ user_id: new ObjectId(user_id) })
             .toArray();
+        // console.log(header);
+
         return { results, totalCaboodles };
     } catch (e) {
         return e;
     }
 };
 
-global.services["DeleteCaboodle"] = async ({ caboodleId }) => {
+global.services["DeleteCaboodle"] = async ({ caboodleId, user_id }, header) => {
     try {
         if (!caboodleId) return { error: "Missing caboodle id" };
+        let authenticatedUser = auth.authenticateRequest(header);
+        if (user_id !== authenticatedUser.user_id)
+            return { error: "Unauthrized user" };
         const results = await global.database["Caboodles"].deleteOne({
             _id: new ObjectId(caboodleId),
         });
@@ -62,13 +71,14 @@ global.services["DeleteCaboodle"] = async ({ caboodleId }) => {
     }
 };
 
-global.services["EditCaboodle"] = async ({
-    caboodleId,
-    name,
-    image,
-    description,
-}) => {
+global.services["EditCaboodle"] = async (
+    { caboodleId, name, image, description, user_id },
+    header
+) => {
     try {
+        let authenticatedUser = auth.authenticateRequest(header);
+        if (user_id !== authenticatedUser.user_id)
+            return { error: "Unauthrized user" };
         const results = await global.database["Caboodles"].updateOne(
             {
                 _id: new ObjectId(caboodleId),
